refactor: exception class 위치 이동
