+++ src/main/java/com/example/twentyfiveframes/exception/CustomException.java
@@ -1,4 +1,4 @@
-package com.example.twentyfiveframes.domain;
+package com.example.twentyfiveframes.exception;
 
 
 import lombok.Getter;
